requirements.txt
fastapi
uvicorn
pydantic
yt-dlp
google-generativeai
httpcore